requirements.txt
Flask
librosa
numpy
psycopg2-binary
soundfile
